ccnl-fwd: fix debug output

+++ src/ccnl-fwd/src/ccnl-dispatch.c
@@ -91,8 +91,8 @@ ccnl_core_RX(struct ccnl_relay_s *relay, int ifndx, unsigned char *data,
             suite = ccnl_pkt2suite(data, datalen, &skip);
 
         if (!ccnl_isSuite(suite)) {
-            DEBUGMSG_CORE(WARNING, "?unknown packet format? ccnl_core_RX ifndx=%d, %d bytes starting with 0x%02x at offset %zd\n",
-                     ifndx, datalen, *data, data - base);
+            DEBUGMSG_CORE(WARNING, "?unknown packet format? ccnl_core_RX ifndx=%d, %d bytes starting with 0x%02x at offset %d\n",
+                     ifndx, datalen, *data, (int)(data - base));
             return;
         }
